// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  DOCTOR: 'DOCTOR',
  STAFF: 'STAFF'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ToothCondition = {
  HEALTHY: 'HEALTHY',
  FILLED: 'FILLED',
  EXTRACTED: 'EXTRACTED',
  NEEDS_TREATMENT: 'NEEDS_TREATMENT',
  CROWN: 'CROWN'
} as const

export type ToothCondition = (typeof ToothCondition)[keyof typeof ToothCondition]


export const AppointmentStatus = {
  CONFIRMED: 'CONFIRMED',
  ATTENDED: 'ATTENDED',
  NO_SHOW: 'NO_SHOW',
  CANCELLED: 'CANCELLED'
} as const

export type AppointmentStatus = (typeof AppointmentStatus)[keyof typeof AppointmentStatus]


export const AppointmentType = {
  CHECKUP: 'CHECKUP',
  FOLLOWUP: 'FOLLOWUP',
  EMERGENCY: 'EMERGENCY'
} as const

export type AppointmentType = (typeof AppointmentType)[keyof typeof AppointmentType]


export const PlanStatus = {
  PROPOSED: 'PROPOSED',
  APPROVED: 'APPROVED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED'
} as const

export type PlanStatus = (typeof PlanStatus)[keyof typeof PlanStatus]


export const InvoiceStatus = {
  PAID: 'PAID',
  PARTIAL: 'PARTIAL',
  OVERDUE: 'OVERDUE'
} as const

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus]


export const PaymentMethod = {
  CASH: 'CASH',
  TRANSFER: 'TRANSFER'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const StockMovementType = {
  ADD: 'ADD',
  CONSUME: 'CONSUME',
  TRANSFER_IN: 'TRANSFER_IN',
  TRANSFER_OUT: 'TRANSFER_OUT'
} as const

export type StockMovementType = (typeof StockMovementType)[keyof typeof StockMovementType]


export const EarningType = {
  PERCENTAGE: 'PERCENTAGE',
  FIXED: 'FIXED',
  MIXED: 'MIXED'
} as const

export type EarningType = (typeof EarningType)[keyof typeof EarningType]


export const ApprovalStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type ApprovalStatus = (typeof ApprovalStatus)[keyof typeof ApprovalStatus]


export const ExpenseCategory = {
  RENT: 'RENT',
  UTILITIES: 'UTILITIES',
  SALARIES: 'SALARIES',
  MAINTENANCE: 'MAINTENANCE',
  MARKETING: 'MARKETING',
  OTHER: 'OTHER'
} as const

export type ExpenseCategory = (typeof ExpenseCategory)[keyof typeof ExpenseCategory]


export const InventoryCategory = {
  MATERIALS: 'MATERIALS',
  TOOLS: 'TOOLS',
  DEVICES: 'DEVICES'
} as const

export type InventoryCategory = (typeof InventoryCategory)[keyof typeof InventoryCategory]